+++ vcf/src/vcf/FoundationKit/OSXSystemPeer.h
@@ -17,7 +17,7 @@ public:
 	OSXSystemPeer();
 	virtual ~OSXSystemPeer();
 
-	virtual unsigned long getTickCount();
+	virtual uint32 getTickCount();
 
 	virtual void sleep( const uint32& milliseconds );
